Improve alignment of text nodes in make-cd
Currently text nodes in a commutative diagram are vertically aligned on their baseline. This is an intuitive and reasonable default, but unfortunately it results in incorrect behavior when combined with snapping; in particular, there are many cases (depending on the physical height of a piece of math text) where drawing an fully horizontal or fully vertical arrow from one math node to another is impossible to achieve with precision, and instead it snaps to something that is at a strange angle. This problem is solved by setting the vertical alignment to be centered.

I am not sure that this is the actual best solution, but it solved the problem for me.