maple plugin: silence compiler warnings
This patch silences the compiler warnings
(gcc (Debian 10.2.1-6) 10.2.1). However
the option `-s` (or `--silence`) given to `make`
in the script `plugins/maple/bin/tm_maple` is kept.
As C compilers are getting more and more verbose,
this choice prevents future (and certainly subtle)
warning-messages to set TeXmacs end-users on the edge
as soon as the _banner stage_.