maple plugin: built machinery: harden
This patch hardens the building machinery of the TeXmacs interface for Maple.

The builder script `plugins/maple/bin/tm_maple` now (1) takes into account any
upgrade of the Maple suite and (2) prepends`$MAPLE_SYS_BIN/system` to the
LIBRARY_PATH passed to the `-L` option and to the RunPATH stored in the
executable when the `glibc` library against which Maple was built is more recent
than the one on the system. This allows on-the-fly upgrades when the builder
script `tm_maple` is not short-cut with a symbolic link in `$TEXMACS_HOME_PATH/bin`.
When `tm_maple` is so short-cut, regular users can upgrade by removing the symbolic
link by hand and restarting a Maple session. Because short-cutting or not the
builder script `tm_maple` has its cons and pros, the choice is let to the user
through a configuration parameter, specifically, the
TEXMACS_HOME_PLUGINS_MAPLE_DO_SYMLINK parameter which is set to `no` by default.
This default choice sets no short-cut and reproduces so the current behaviour.
The user can set this parameter to `yes` (or anything else than `no`)
in the configuration file `$TEXMACS_HOME_PATH/plugins/maple/tm_maple.conf`
to allow short-cutting at building time. Note that the script sources the
configuration file after that the `maple` to consider is set up (shell variable
`TEXMACS_MAPLE_BIN`) so that the user can specify their favourite `maple`
in the configuration file.

From a technical point, upgrades of the Maple suite are guessed
by using as `STAMP` a file installed or upgraded for sure by
Maple Installers or Upgraders. The selection of the right `STAMP`
is based on my `SingleUser license type` installation. We can be
confident that the license file `$MAPLE/license/license.dat` will be
updated. But the license must be "only" updated when `major` upgrade
occurs. For `minor` (and `patch` ?) upgrades we may consider the
`$MAPLE/license/version.txt`. But, contrary to the license file,
the version file is not mandatory to run Maple. This is why
the latter is selected as first choice, and the former as second
choice. If both are not present or not at their default place,
an ad hoc `STAMP` file is created with the `compilation date` extracted
from the `kernelopts(version)` string. For other type license, I have
no idea and no access. In any case (1) one can increase the list of `STAMP`
candidates, (2) the ad hoc `STAMP` will make the trick, and (3) one can
introduced a dedicated configuration parameter.